+++ Cesium3DTilesSelection/test/TestTilesetContentManager.cpp
@@ -372,7 +372,7 @@ TEST_CASE("Test tile state machine") {
 
       // ContentLoaded -> Done
       // update tile content to move from ContentLoaded -> Done
-      pManager->updateTileContent(tile, 0.0, options);
+      pManager->updateTileContent(tile, options);
       CHECK(tile.getState() == TileLoadState::Done);
       CHECK(tile.getChildren().size() == 1);
       CHECK(tile.getChildren().front().getContent().isEmptyContent());
@@ -473,7 +473,7 @@ TEST_CASE("Test tile state machine") {
 
     // FailedTemporarily -> FailedTemporarily
     // tile is failed temporarily but the loader can still add children to it
-    pManager->updateTileContent(tile, 0.0, options);
+    pManager->updateTileContent(tile, options);
     CHECK(pManager->getNumberOfTilesLoading() == 0);
     CHECK(tile.getChildren().size() == 1);
     CHECK(tile.getChildren().front().isEmptyContent());
@@ -548,7 +548,7 @@ TEST_CASE("Test tile state machine") {
 
     // Failed -> Failed
     // tile is failed but the loader can still add children to it
-    pManager->updateTileContent(tile, 0.0, options);
+    pManager->updateTileContent(tile, options);
     CHECK(pManager->getNumberOfTilesLoading() == 0);
     CHECK(tile.getChildren().size() == 1);
     CHECK(tile.getChildren().front().isEmptyContent());
@@ -648,7 +648,7 @@ TEST_CASE("Test tile state machine") {
     CHECK(upsampledTile.getState() == TileLoadState::Unloaded);
 
     // parent moves from ContentLoaded -> Done
-    pManager->updateTileContent(tile, 0.0, options);
+    pManager->updateTileContent(tile, options);
     CHECK(tile.getState() == TileLoadState::Done);
     CHECK(tile.getChildren().size() == 1);
     CHECK(&tile.getChildren().back() == &upsampledTile);